add test for deprecation
