+++ tests/deprecated_api/test_remove_1-8.py
@@ -12,14 +12,29 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 """Test deprecated functionality which will be removed in v1.8.0."""
+import os
+import sys
+from unittest import mock
+from unittest.mock import Mock
 
 import pytest
 import torch
 
+from pytorch_lightning.utilities import AllGatherGrad
 from pytorch_lightning.utilities.apply_func import move_data_to_device
 from pytorch_lightning.utilities.enums import DeviceType, DistributedType
 from pytorch_lightning.utilities.imports import _TORCHTEXT_LEGACY
+from tests.helpers.runif import RunIf
 from tests.helpers.torchtext_utils import get_dummy_torchtext_data_iterator
+
+
+def setup_ddp(rank, world_size):
+    """Setup ddp enviroment."""
+    os.environ["MASTER_ADDR"] = "localhost"
+    os.environ["MASTER_PORT"] = "8088"
+
+    if torch.distributed.is_available() and sys.platform not in ("win32", "cygwin"):
+        torch.distributed.init_process_group("gloo", rank=rank, world_size=world_size)
 
 
 def test_v1_8_0_deprecated_distributed_type_enum():
@@ -41,3 +56,13 @@ def test_v1_8_0_deprecated_torchtext_batch():
         data_iterator, _ = get_dummy_torchtext_data_iterator(num_samples=3, batch_size=3)
         batch = next(iter(data_iterator))
         _ = move_data_to_device(batch=batch, device=torch.device("cpu"))
+
+
+@RunIf(skip_windows=True)
+@mock.patch("torch.distributed.all_gather")
+@mock.patch("torch.distributed.get_world_size", return_value=1)
+def test_v1_8_0_deprecated_all_gather_grad(*_):
+    tensor1 = torch.ones(1, requires_grad=True)
+
+    with pytest.deprecated_call(match="`AllGatherGrad` has been deprecated in v1.6"):
+        AllGatherGrad.apply(tensor1)
